--- src/main/java/kr/hhplus/be/server/api/order/response/OrderCreateResponse.java
@@ -1,6 +1,0 @@
-package kr.hhplus.be.server.api.order.response;
-
-public record OrderCreateResponse(
-	Long orderId
-) {
-}
