--- src/main/java/kr/hhplus/be/server/api/product/response/BestSellerReadAllResponse.java
@@ -1,12 +1,0 @@
-package kr.hhplus.be.server.api.product.response;
-
-import java.util.List;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-
-@Schema(name = "베스트셀러 목록 조회 응답")
-public record BestSellerReadAllResponse(
-	@Schema(name = "베스트셀러 목록", description = "베스트셀러 목록")
-	List<BestSellerSimpleInfo> bestSellers
-) {
-}
