+++ src/main/java/kr/hhplus/be/server/interfaces/api/order/response/OrderCreateResponse.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.interfaces.api.order.response;
+
+public record OrderCreateResponse(
+	Long orderId
+) {
+}
