+++ src/test/java/kr/hhplus/be/server/api/product/ProductControllerE2ETest.java
@@ -1,8 +1,6 @@
 package kr.hhplus.be.server.api.product;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import org.junit.jupiter.api.DisplayName;
@@ -11,7 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.web.servlet.MockMvc;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
+import kr.hhplus.be.server.interfaces.api.product.ProductController;
 
 @WebMvcTest(ProductController.class)
 class ProductControllerE2ETest {
